refactor: 패키지 분리
humbledude.carracing.domin: 도메인 객체
humbledude.carracing.view: UI 관련 객체
