+++ src/main/java/humbledude/carracing/domain/Car.java
@@ -1,4 +1,4 @@
-package humbledude.carracing;
+package humbledude.carracing.domain;
 
 public class Car {
