+++ src/main/java/humbledude/carracing/domain/FuelGenerator.java
@@ -1,4 +1,4 @@
-package humbledude.carracing;
+package humbledude.carracing.domain;
 
 public interface FuelGenerator {
     int generate();
