+++ src/test/java/humbledude/carracing/domain/CarTest.java
@@ -1,6 +1,6 @@
-package humbledude.carracing;
+package humbledude.carracing.domain;
 
-import org.junit.jupiter.api.Test;
+import humbledude.carracing.domain.Car;
 import org.junit.jupiter.params.ParameterizedTest;
 import org.junit.jupiter.params.provider.CsvSource;
 import org.junit.jupiter.params.provider.ValueSource;
